Add cross-platform support

+++ src/audio/synthInternals.h
@@ -161,7 +161,7 @@ typedef struct {
     s32                         sample;
     s32                         lastsam;
     s32                         first;
-    s32                         memin; 
+    intptr_t                    memin; 
 } ALLoadFilter;
 
 void    alLoadNew(ALLoadFilter *f, ALDMANew dma, ALHeap *hp);
@@ -254,7 +254,7 @@ s32     alResampleParam(void *f, s32 paramID, void *param);
 
 typedef struct ALSave_s {
     ALFilter            filter;
-    s32	       		dramout;
+    intptr_t	       	dramout;
     s32                 first;
 } ALSave;
